refactor: refacts models package to DTOs
