+++ src/main/java/com/cube/user/dtos/request/RequestUser.java
@@ -1,4 +1,4 @@
-package com.cube.user.models.request;
+package com.cube.user.dtos.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
