+++ src/main/java/com/cube/user/dtos/response/ResponseLogin.java
@@ -1,3 +1,3 @@
-package com.cube.user.models.response;
+package com.cube.user.dtos.response;
 
 public record ResponseLogin (String token) { }
