+++ src/main/java/com/cube/user/dtos/response/ResponseValidate.java
@@ -1,3 +1,3 @@
-package com.cube.user.models.response;
+package com.cube.user.dtos.response;
 
 public record ResponseValidate (String username) {}
