+++ src/main/java/com/cube/user/exceptions/UnauthorizedException.java
@@ -1,6 +1,6 @@
 package com.cube.user.exceptions;
 
-import com.cube.user.models.internal.ExceptionCode;
+import com.cube.user.dtos.internal.ExceptionCode;
 
 public class UnauthorizedException extends BusinessException {
     public UnauthorizedException(ExceptionCode exceptionCode) {
