+++ src/test/java/com/cube/user/services/AuthServiceTest.java
@@ -3,11 +3,11 @@ package com.cube.user.services;
 import com.cube.user.exceptions.ConflitException;
 import com.cube.user.factory.TokenFactory;
 import com.cube.user.factory.UserFactory;
-import com.cube.user.models.internal.InternalUser;
-import com.cube.user.models.request.RequestLogin;
-import com.cube.user.models.request.RequestUser;
-import com.cube.user.models.request.RequestValidate;
-import com.cube.user.models.response.ResponseUser;
+import com.cube.user.models.InternalUser;
+import com.cube.user.dtos.request.RequestLogin;
+import com.cube.user.dtos.request.RequestUser;
+import com.cube.user.dtos.internal.RequestValidate;
+import com.cube.user.dtos.response.ResponseUser;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
