+++ src/test/java/com/cube/user/services/UserServiceTests.java
@@ -3,9 +3,9 @@ package com.cube.user.services;
 import com.cube.user.exceptions.NotFoundException;
 import com.cube.user.factory.UserFactory;
 import com.cube.user.mappers.UserMapper;
-import com.cube.user.models.internal.InternalUser;
-import com.cube.user.models.request.RequestUser;
-import com.cube.user.models.response.ResponseUser;
+import com.cube.user.models.InternalUser;
+import com.cube.user.dtos.request.RequestUser;
+import com.cube.user.dtos.response.ResponseUser;
 import com.cube.user.repositories.UserRepository;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
